requirements.txt
Pillow
prettytable~=3.10.0